+++ apm-sniffer/apm-sdk-plugin/rabbitmq-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/rabbitmq/RabbitMQConsumerInterceptor.java
@@ -18,61 +18,29 @@
 
 package org.apache.skywalking.apm.plugin.rabbitmq;
 
-import com.rabbitmq.client.AMQP;
-import com.rabbitmq.client.Envelope;
-import org.apache.skywalking.apm.agent.core.context.CarrierItem;
-import org.apache.skywalking.apm.agent.core.context.ContextCarrier;
-import org.apache.skywalking.apm.agent.core.context.ContextManager;
-import org.apache.skywalking.apm.agent.core.context.tag.Tags;
-import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
-import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
+import com.rabbitmq.client.Consumer;
+import java.lang.reflect.Method;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
-
-import java.lang.reflect.Method;
 
 public class RabbitMQConsumerInterceptor implements InstanceMethodsAroundInterceptor {
-    public static final String OPERATE_NAME_PREFIX = "RabbitMQ/";
-    public static final String CONSUMER_OPERATE_NAME_SUFFIX = "/Consumer";
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
         MethodInterceptResult result) throws Throwable {
-        ContextCarrier contextCarrier = new ContextCarrier();
-        String url = (String) objInst.getSkyWalkingDynamicField();
-        Envelope envelope = (Envelope) allArguments[1];
-        AMQP.BasicProperties properties = (AMQP.BasicProperties) allArguments[2];
-        AbstractSpan activeSpan = ContextManager.createEntrySpan(OPERATE_NAME_PREFIX + "Topic/" + envelope.getExchange() + "Queue/" + envelope
-            .getRoutingKey() + CONSUMER_OPERATE_NAME_SUFFIX, null).start(System.currentTimeMillis());
-        Tags.MQ_BROKER.set(activeSpan, url);
-        Tags.MQ_TOPIC.set(activeSpan, envelope.getExchange());
-        Tags.MQ_QUEUE.set(activeSpan, envelope.getRoutingKey());
-        activeSpan.setComponent(ComponentsDefine.RABBITMQ_CONSUMER);
-        SpanLayer.asMQ(activeSpan);
-        CarrierItem next = contextCarrier.items();
-        while (next.hasNext()) {
-            next = next.next();
-            if (properties.getHeaders() != null && properties.getHeaders().get(next.getHeadKey()) != null) {
-                next.setHeadValue(properties.getHeaders().get(next.getHeadKey()).toString());
-            }
-        }
-        ContextManager.extract(contextCarrier);
-
+        Consumer consumer = (Consumer) allArguments[6];
+        allArguments[6] = new TracerConsumer(consumer, (String) objInst.getSkyWalkingDynamicField());
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
         Object ret) throws Throwable {
-        ContextManager.stopSpan();
         return ret;
-
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
         Class<?>[] argumentsTypes, Throwable t) {
-        ContextManager.activeSpan().log(t);
     }
 }